refactor: move custom_analyzer to pdf_extract.py
